requirements.txt
printy==2.1.1
python-dotenv==0.15.0
selenium==3.141.0
urllib3==1.26.3
